lightrag/api/requirements.txt
aiofiles
ascii_colors
docx
python-pptx
openpyxl
fastapi
graspologic
numpy
pipmaster
python-dotenv
python-multipart
tenacity
tiktoken
uvicorn
